services/aggregator-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
boto3
APScheduler
python-dotenv
psycopg2-binary
sqlalchemy-redshift
requests